In Commodity, eliminated call to sqloxx::SQLStatement::bind<unsigned char>(...), since unsigned char is no longer supported by this function.
git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2422 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98

+++ src/commodity.cpp
@@ -217,7 +217,10 @@ void Commodity::process_saving_statement(SQLStatement& statement)
 	statement.bind(":precision", value(m_data->precision));
 	Decimal m = value(m_data->multiplier_to_base);
 	statement.bind(":multiplier_to_base_intval", m.intval());
-	statement.bind(":multiplier_to_base_places", m.places());
+	statement.bind
+	(	":multiplier_to_base_places",
+		numeric_cast<int>(m.places())
+	);
 	statement.step_final();	
 	return;
 }
